Auth fails early if server has an error on it
